requirements.txt
pyjwt==2.8.0
Django==5.1.4
django-ninja==1.3.0
